console: Remove old side navigation
