+++ pkg/webui/components/header/index.js
@@ -15,9 +15,8 @@
 import React, { useRef } from 'react'
 import classnames from 'classnames'
 
-import { Breadcrumbs } from '@ttn-lw/components/breadcrumbs/breadcrumbs'
 import Button from '@ttn-lw/components/button'
-import ProfileDropdown from '@ttn-lw/components/profile-dropdown-v2'
+import ProfileDropdown from '@ttn-lw/components/profile-dropdown'
 import Link from '@ttn-lw/components/link'
 
 import PropTypes from '@ttn-lw/lib/prop-types'
