--- pkg/webui/components/navigation/side-v2/item/item.styl
@@ -1,87 +1,0 @@
-// Copyright © 2019 The Things Network Foundation, The Things Industries B.V.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-.item
-  &:hover
-    .fly-out-list
-      display: block
-
-.button
-  reset-button()
-  width: 100%
-  transition: 0.2s background-color ease-in, 0.2s color ease-in
-  text-decoration: none
-  display: flex
-  align-items: center
-  position: relative
-  gap: $cs.xs !important
-
-  span
-    &:not(:first-child)
-      color: $c.grey-700 !important
-
-  &:hover&:not(&-active)
-    background: none !important
-    color: $tc-deep-gray !important
-    text-decoration: none
-
-    span
-      &:not(:first-child)
-        color: $c.grey-900 !important
-
-    .icon
-      color: inherit
-
-  &-active
-    background: $c.tts-primary-150
-    color: $c.grey-900 !important
-
-    &:hover
-      background: $c.tts-primary-150 !important
-      color: $c.grey-900 !important
-
-      span
-        &:not(:first-child)
-          color: $c.grey-900 !important
-
-      .icon
-        color: inherit
-
-    .icon
-      color: inherit
-
-.message
-  vertical-align: middle
-
-.expand-icon
-  position: absolute
-  right: $cs.xs
-  top: 50%
-  transform: translateY(-50%)
-  transition: transform $ad.m ease-in
-
-  &-open
-    transform: translateY(-50%) rotate(180deg)
-
-.fly-out-list-container
-  top: 0
-  left: 0
-  position: absolute
-  width: 180%
-  height: 13rem
-
-.fly-out-list
-  display: none
-  left: 5rem !important
-  top: 0 !important
