--- pkg/webui/components/navigation/side-v2/side.styl
@@ -1,91 +1,0 @@
-// Copyright © 2019 The Things Network Foundation, The Things Industries B.V.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-.container
-  border-normal(right)
-  display: flex
-  flex-direction: column
-  z-index: $zi.nav
-  +media-query-min($bp.s)
-    left: 0
-    top: 0
-    position: absolute
-    height: 100%
-  +media-query($bp.s)
-    height: auto
-    position: sticky
-    top: $header-height-mobile
-    z-index: $zi.nav
-    +media-query-height(500px)
-      position: static
-
-.navigation
-  box-sizing: border-box
-  max-height: 100%
-  width: 100%
-
-  &-minimized
-    overflow: visible
-    display: flex
-    flex-direction: column
-    justify-content: space-between
-    +media-query-min($bp.s)
-      width: $sidebar-width-minimized
-      display: inline-flex
-
-      .header
-        padding: 0
-        width: 100%
-        display: flex
-        height: $breadcrumbs-bar-height
-        justify-content: center
-
-        .message
-          display: none
-
-        .icon
-          margin-right: 0
-
-      .navigation-list
-        overflow: visible
-
-      .drawer
-        overflow: visible
-
-.drawer
-  background-color: white
-  z-index: $zi.nav
-  overflow: auto
-  +media-query($bp.s)
-    sidebar-transition(left, box-shadow)
-    position: fixed
-    top: $header-height-mobile
-    left: -60vw
-    bottom: 0
-    width: 60vw
-    max-width: 20rem
-
-    &-open
-      box-shadow: 0 0 13rem 2rem rgba(0,0,0,.4)
-      left: 0
-
-.message
-  display: block
-  white-space: nowrap
-  overflow: hidden
-  text-overflow: ellipsis
-
-.scroll-lock
-  +media-query($bp.s)
-    overflow: hidden
