+++ pkg/webui/components/navigation/side/list/list.styl
@@ -16,7 +16,8 @@
   list-style-type: none
   padding: 0
   margin: 0
-  overflow: hidden
+  overflow: visible
+  font-size: $fs.m
 
   & > li:not(:last-child)
     margin-bottom: 0
@@ -28,18 +29,18 @@
     transition: height $ad.m ease-in-out
 
   &-nested > li
-    height: 0
-    transition: visibility $ad.m, height $ad.m ease-in-out
+    height: 0px
+    transition: height $ad.m ease-in-out
 
   &-nested:not(.list-expanded) > li
     visibility: hidden
 
   &-expanded > li
-    height: 3.1rem
+    height: 2.4rem
 
   &-expanded, &-nested
     & li a
-      padding-left: $cs.l + $cs.l
+      padding-left: $cs.l + $cs.xs
 
 .icon
   margin-right: $cs.xs
