--- pkg/webui/components/profile-dropdown-v2/profile-dropdown-v2.styl
@@ -1,28 +1,0 @@
-// Copyright © 2023 The Things Network Foundation, The Things Industries B.V.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-.container
-  padding: 0 0.25rem 0 0.5rem
-  position: relative
-  gap: 0
-
-.brand-logo
-  height: 1.14rem
-
-  +media-query($bp.s)
-    display: none
-
-.profile-picture
-  height: 1.5rem !important
-  width: 1.5rem !important
