+++ pkg/webui/components/sidebar/side-footer/side-footer.styl
@@ -26,4 +26,7 @@
     right: -79px !important
 
   &-dropdown
-    width: 18rem
+    width: 18.2rem
+
+  &-cluster-dropdown
+    left: -207px !important
