+++ pkg/webui/console/containers/side-bar/navigation/general-side-navigation.js
@@ -15,7 +15,7 @@
 import React, { useContext } from 'react'
 import { useSelector } from 'react-redux'
 
-import SideNavigation from '@ttn-lw/components/navigation/side-v2'
+import SideNavigation from '@ttn-lw/components/navigation/side'
 import SectionLabel from '@ttn-lw/components/sidebar/section-label'
 
 import sharedMessages from '@ttn-lw/lib/shared-messages'
