+++ pkg/webui/console/containers/side-bar/side-bar.styl
@@ -18,6 +18,7 @@
   font-size: $fs.s
   height: calc(100vh - 1.5rem)
   min-width: 20rem
+  width: 20rem
   sidebar-transition(min-width)
   +media-query($bp.m)
     sidebar-transition(left, box-shadow)
