+++ pkg/webui/console/views/application/index.js
@@ -16,12 +16,8 @@ import React, { useEffect } from 'react'
 import { Routes, Route, useParams } from 'react-router-dom'
 import { useDispatch, useSelector } from 'react-redux'
 
-import applicationIcon from '@assets/misc/application.svg'
-
 import { useBreadcrumbs } from '@ttn-lw/components/breadcrumbs/context'
-import SideNavigation from '@ttn-lw/components/navigation/side'
 import Breadcrumb from '@ttn-lw/components/breadcrumbs/breadcrumb'
-import Breadcrumbs from '@ttn-lw/components/breadcrumbs'
 
 import IntlHelmet from '@ttn-lw/lib/components/intl-helmet'
 import RequireRequest from '@ttn-lw/lib/components/require-request'
@@ -41,22 +37,9 @@ import ApplicationIntegrationsMqtt from '@console/views/application-integrations
 import ApplicationIntegrationsLoRaCloud from '@console/views/application-integrations-lora-cloud'
 import Devices from '@console/views/devices'
 
-import sharedMessages from '@ttn-lw/lib/shared-messages'
 import { selectApplicationSiteName } from '@ttn-lw/lib/selectors/env'
 
-import {
-  mayViewApplicationInfo,
-  mayViewApplicationEvents,
-  maySetApplicationPayloadFormatters,
-  mayViewApplicationDevices,
-  mayCreateOrEditApplicationIntegrations,
-  mayEditBasicApplicationInfo,
-  mayViewOrEditApplicationApiKeys,
-  mayViewOrEditApplicationCollaborators,
-  mayViewOrEditApplicationPackages,
-  mayAddPubSubIntegrations,
-  mayViewApplications,
-} from '@console/lib/feature-checks'
+import { mayAddPubSubIntegrations, mayViewApplications } from '@console/lib/feature-checks'
 
 import {
   getApplication,
@@ -65,10 +48,7 @@ import {
 } from '@console/store/actions/applications'
 import { getAsConfiguration } from '@console/store/actions/application-server'
 
-import {
-  selectApplicationRights,
-  selectSelectedApplication,
-} from '@console/store/selectors/applications'
+import { selectSelectedApplication } from '@console/store/selectors/applications'
 import {
   selectMqttProviderDisabled,
   selectNatsProviderDisabled,
@@ -102,7 +82,6 @@ const ApplicationInner = () => {
   const { appId } = useParams()
   const application = useSelector(selectSelectedApplication)
   const name = application.name || appId
-  const rights = useSelector(selectApplicationRights)
   const siteName = selectApplicationSiteName()
   const natsDisabled = useSelector(selectNatsProviderDisabled)
   const mqttDisabled = useSelector(selectMqttProviderDisabled)
@@ -115,85 +94,7 @@ const ApplicationInner = () => {
 
   return (
     <>
-      <Breadcrumbs />
       <IntlHelmet titleTemplate={`%s - ${name} - ${siteName}`} />
-      <SideNavigation
-        header={{
-          icon: applicationIcon,
-          iconAlt: sharedMessages.application,
-          title: name,
-          to: '',
-        }}
-      >
-        {mayViewApplicationInfo.check(rights) && (
-          <SideNavigation.Item title={sharedMessages.overview} path="" icon="overview" exact />
-        )}
-        {mayViewApplicationDevices.check(rights) && (
-          <SideNavigation.Item title={sharedMessages.devices} path="devices" icon="devices" />
-        )}
-        {mayViewApplicationEvents.check(rights) && (
-          <SideNavigation.Item title={sharedMessages.liveData} path="data" icon="data" />
-        )}
-        {maySetApplicationPayloadFormatters.check(rights) && (
-          <SideNavigation.Item title={sharedMessages.payloadFormatters} icon="code">
-            <SideNavigation.Item
-              title={sharedMessages.uplink}
-              path="payload-formatters/uplink"
-              icon="uplink"
-            />
-            <SideNavigation.Item
-              title={sharedMessages.downlink}
-              path="payload-formatters/downlink"
-              icon="downlink"
-            />
-          </SideNavigation.Item>
-        )}
-        {mayCreateOrEditApplicationIntegrations.check(rights) && (
-          <SideNavigation.Item title={sharedMessages.integrations} icon="integration">
-            <SideNavigation.Item
-              title={sharedMessages.mqtt}
-              path="integrations/mqtt"
-              icon="extension"
-            />
-            <SideNavigation.Item
-              title={sharedMessages.webhooks}
-              path="integrations/webhooks"
-              icon="extension"
-            />
-            {mayAddPubSubIntegrations.check(natsDisabled, mqttDisabled) && (
-              <SideNavigation.Item
-                title={sharedMessages.pubsubs}
-                path="integrations/pubsubs"
-                icon="extension"
-              />
-            )}
-            {mayViewOrEditApplicationPackages.check(rights) && (
-              <SideNavigation.Item
-                title={sharedMessages.loraCloud}
-                path="integrations/lora-cloud"
-                icon="extension"
-              />
-            )}
-          </SideNavigation.Item>
-        )}
-        {mayViewOrEditApplicationCollaborators.check(rights) && (
-          <SideNavigation.Item
-            title={sharedMessages.collaborators}
-            path="collaborators"
-            icon="organization"
-          />
-        )}
-        {mayViewOrEditApplicationApiKeys.check(rights) && (
-          <SideNavigation.Item title={sharedMessages.apiKeys} path="api-keys" icon="api_keys" />
-        )}
-        {mayEditBasicApplicationInfo.check(rights) && (
-          <SideNavigation.Item
-            title={sharedMessages.generalSettings}
-            path="general-settings"
-            icon="general_settings"
-          />
-        )}
-      </SideNavigation>
       <Routes>
         <Route index Component={ApplicationOverview} />
         <Route path="general-settings" Component={ApplicationGeneralSettings} />
