+++ pkg/webui/console/views/applications-list/index.js
@@ -15,21 +15,30 @@
 import React from 'react'
 import { Row, Col, Container } from 'react-grid-system'
 
+import Breadcrumb from '@ttn-lw/components/breadcrumbs/breadcrumb'
+import { useBreadcrumbs } from '@ttn-lw/components/breadcrumbs/context'
+
 import IntlHelmet from '@ttn-lw/lib/components/intl-helmet'
 
 import ApplicationsTable from '@console/containers/applications-table'
 
 import sharedMessages from '@ttn-lw/lib/shared-messages'
 
-const ApplicationsList = () => (
-  <Container>
-    <Row>
-      <IntlHelmet title={sharedMessages.applications} />
-      <Col>
-        <ApplicationsTable />
-      </Col>
-    </Row>
-  </Container>
-)
+const ApplicationsList = () => {
+  useBreadcrumbs('apps.list', <Breadcrumb path="/applications" content={sharedMessages.list} />)
+
+  return (
+    <>
+      <Container>
+        <Row>
+          <IntlHelmet title={sharedMessages.applications} />
+          <Col>
+            <ApplicationsTable />
+          </Col>
+        </Row>
+      </Container>
+    </>
+  )
+}
 
 export default ApplicationsList
