+++ pkg/webui/console/views/gateways-list/index.js
@@ -15,21 +15,28 @@
 import React from 'react'
 import { Container, Row, Col } from 'react-grid-system'
 
+import Breadcrumb from '@ttn-lw/components/breadcrumbs/breadcrumb'
+import { useBreadcrumbs } from '@ttn-lw/components/breadcrumbs/context'
+
 import IntlHelmet from '@ttn-lw/lib/components/intl-helmet'
 
 import GatewaysTable from '@console/containers/gateways-table'
 
 import sharedMessages from '@ttn-lw/lib/shared-messages'
 
-const GatewaysList = () => (
-  <Container>
-    <Row>
-      <IntlHelmet title={sharedMessages.gateways} />
-      <Col>
-        <GatewaysTable />
-      </Col>
-    </Row>
-  </Container>
-)
+const GatewaysList = () => {
+  useBreadcrumbs('gtws.list', <Breadcrumb path="/gateways" content={sharedMessages.list} />)
+
+  return (
+    <Container>
+      <Row>
+        <IntlHelmet title={sharedMessages.gateways} />
+        <Col>
+          <GatewaysTable />
+        </Col>
+      </Row>
+    </Container>
+  )
+}
 
 export default GatewaysList
